src/orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alicepi-proto",
    "pyzmq==25.1.1",
    "requests==2.31.0",
    "websockets==12.0",
    "pydub==0.25.1",
    "numpy",
]

[tool.uv.sources]
alicepi-proto = { path = "../proto_defs" }
